backend/requirements.txt

flask==3.0.3
flask-cors==4.0.1
flask-caching==2.3.0
flask-limiter==3.7.0
flasgger==0.9.7.1                 

celery[redis]==5.5.2
redis==5.0.8
kombu==5.5.2
flower==2.0.1                     

playwright==1.45.1
selenium==4.23.1
webdriver-manager==4.0.2
scrapy==2.11.2
requests==2.32.3
httpx==0.27.0
aiohttp==3.10.2
beautifulsoup4==4.12.3
cssselect==1.2.0
parsel==1.9.1
fake-useragent==1.5.1
cloudscraper==1.2.71              
stem==1.8.2                        
pyppeteer==2.0.0                   
arsenic==21.8                      
mechanize==0.4.10                  
curl-cffi==0.7.1

pandas==2.2.3
polars==1.3.0
numpy==2.3.5
pyarrow==23.0.1
dask==2026.3.0
pyspark==4.1.1                   # Out-of-core DataFrames

sqlalchemy==2.0.32
alembic==1.13.2
pymongo==4.8.0
motor==3.5.0
aioredis==2.0.1
elasticsearch==8.14.0
clickhouse-driver==0.2.7


google-cloud-bigquery==3.40.1
google-cloud-storage==3.10.1
google-cloud-pubsub==2.36.0
boto3==1.42.78                      # AWS S3/Glue
azure-storage-blob==12.22.0

matplotlib==3.9.2
seaborn==0.13.2
plotly==5.23.0
dash==4.1.0
dash-bootstrap-components==2.0.4
bokeh==3.9.0

tensorflow==2.21.0
keras==3.13.2
scikit-learn==1.8.0
xgboost==2.1.1
lightgbm==4.5.0
statsmodels==0.14.6
prophet==1.3.0                  
optuna==3.6.1                     
mlflow==3.10.1                    
torch==2.11.0                      
transformers==5.4.0             
sentence-transformers==5.3.0      
spacy==3.8.14                
nltk==3.9.1
openai==2.30.0

cython==3.0.11
numba==0.62.1
numexpr==2.10.1
bottleneck==1.4.0

great-expectations==1.15.1
pandera==0.20.3
evidently==0.4.33                 
ydata-profiling==4.18.1            

openpyxl==3.1.5
xlsxwriter==3.2.0
reportlab==4.2.2                   
weasyprint==62.3                 
python-docx==1.1.2
jinja2==3.1.6


pydantic==2.8.2
pydantic-settings==2.4.0
marshmallow==3.22.0
orjson==3.10.7

python-dotenv==1.0.1
loguru==0.7.2
rich==13.7.1                       
typer==0.12.3                      
schedule==1.2.2
apscheduler==3.10.4
pytz==2026.1.post1
arrow==1.3.0                       
tqdm==4.66.5
tenacity==9.0.0                    
cachetools==7.0.5
psutil==6.0.0                      
watchdog==6.0.0

cryptography==43.0.0
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
httpretty==1.1.4
responses==0.25.3
factory-boy==3.3.1

black==24.8.0
ruff==0.5.7
mypy==1.11.1
pre-commit==3.8.0
ipython==8.26.0
jupyter==1.0.0